feat: Introduce AI news response structure
Refactor the `AINewsItem` type to be part of a new `AINewsResponse` interface. This includes adding a `type` field to `AINewsItem` and introducing `AIFallbackHighlight` and `AINewsResponse` types. The `aiNews` state in `App.tsx` is updated to `aiNewsData` to reflect this change in data structure. Cache keys are also updated to version 17.